+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Highlight located tweak.
 - Auto generated documentation.
 
+### Fixed
+- Pan gesture conflict in floating panel.
+
 ## [1.0.1] - 2017-06-20
 ### Added
 - More documentation for public API.